requirements.txt
fastapi
uvicorn
pydantic
jinja2
sqlalchemy
reportlab
python-jose
passlib[bcrypt]
python-multipart
